% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate Data}
\usage{
aggregate_data(data)
}
\arguments{
\item{data}{The input data frame.}
}
\value{
A new data frame with the aggregated data.
}
\description{
This function aggregates the given data by grouping it based on the chunk, gap, and itemtype columns.
It then calculates various summary statistics including the total count, the count of correct values,
and the proportion of correct values.
}
\examples{
tar_load(exp1_data)
aggregate_data(exp1_data)

}
