% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boot_est}
\alias{boot_est}
\title{Perform bootstrapped estimation}
\usage{
boot_est(data, start, exclude_sp1 = TRUE, growth = "asy", ...)
}
\arguments{
\item{data}{The input data for estimation.}

\item{start}{The starting values for estimation. If not provided, default values will be used.}

\item{exclude_sp1}{Logical value indicating whether to exclude the first item in the data. Default is TRUE.}

\item{growth}{The growth type for estimation. Default is "asy".}

\item{...}{Additional arguments to be passed to the estimation function.}
}
\value{
A data frame containing the estimated model parameters.
}
\description{
This function performs bootstrapped estimation using the provided data and parameters.
It uses the serial recall model for estimation
}
\examples{
data <- read.csv("data.csv")
boot_est(data)

}
