% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{extract_object_from_rdata}
\alias{extract_object_from_rdata}
\title{Get data object from a file}
\usage{
extract_object_from_rdata(path, object_name = "data_an")
}
\arguments{
\item{path}{The path to the file containing the Rdata file}

\item{object_name}{The name of the object to extract from the Rdata file}
}
\value{
The loaded data object.
}
\description{
This function safely loads the environment from an Rdata file and returns
an object from it.
}
