% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{gen_boot_dataset}
\alias{gen_boot_dataset}
\title{Generate a bootstrapped dataset}
\usage{
gen_boot_dataset(data)
}
\arguments{
\item{data}{The input dataset}
}
\value{
A bootstrapped dataset with aggregated summary statistics
}
\description{
This function takes in a dataset and generates a bootstrapped dataset by randomly sampling
unique IDs with replacement. For each ID and condition, it calculates the proportion of correct
responses and then generates a new dataset by sampling from a binomial distribution with the
calculated proportion of correct responses.
}
\examples{
data <- read.csv("data.csv")
boot_data <- gen_boot_dataset(data)

}
