% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{get_data}
\alias{get_data}
\title{get_data function}
\usage{
get_data(path, ...)
}
\arguments{
\item{path}{The file path to the RData file.}
}
\value{
The preprocessed data.
}
\description{
This function takes a file path as input and returns preprocessed data.
}
\seealso{
extract_object_from_rdata preprocess_data
}
