% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{Logit Transformation}
\usage{
logit(x, lb = 0, ub = 1)
}
\arguments{
\item{x}{The variable to be transformed.}

\item{lb}{The lower bound of the variable (default is 0).}

\item{ub}{The upper bound of the variable (default is 1).}
}
\value{
The logit-transformed variable.
}
\description{
This function performs a logit transformation on a given variable.
}
\examples{
logit(0.5) # returns 0
logit(0.25, lb = 0, ub = 0.5) # returns 0
logit(0.75, lb = 0.5, ub = 1) # returns 0

}
