% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{overall_deviance}
\alias{overall_deviance}
\title{Calculate the overall deviance}
\usage{
overall_deviance(params, data, by = c("chunk", "gap"), ..., priors = list())
}
\arguments{
\item{params}{A vector of parameters.}

\item{data}{A data frame containing the data.}

\item{by}{The column(s) to split the data by.}
}
\value{
The overall deviance.
}
\description{
This function calculates the overall deviance for a given set of parameters and data.
It splits the dataset by the given column and calculates the deviance for each subset.
The overall deviance is the sum of the deviances for each subset.
}
