% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_bootstrap_results}
\alias{plot_bootstrap_results}
\title{Plot Bootstrap Results}
\usage{
plot_bootstrap_results(data)
}
\arguments{
\item{data}{The dataset containing the bootstrap results.}
}
\value{
A ggplot object displaying the histogram of the bootstrap results for each parameter.
}
\description{
This function plots the bootstrap results for a given dataset.
}
\examples{
tar_load(exp1_data)
data <- replicate(10, boot_est(exp1_data), simplify = FALSE)
data <- do.call(rbind, data)
plot_bootstrap_results(data)

}
