% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_linear_rv_recovery}
\alias{plot_linear_rv_recovery}
\title{Plot Linear RV Recovery}
\usage{
plot_linear_rv_recovery(r, t, title = "Uniform distribution")
}
\arguments{
\item{r}{A vector representing the recovery rates sampled from some distribution.}

\item{t}{A vector representing the time values.}

\item{title}{The title of the plot (default is "Uniform distribution").}
}
\value{
A ggplot object representing the plot of linear RV recovery over time.
}
\description{
This function plots the recovery of a linear random variable (RV) over time.
}
\details{
Average recovery is given as a red line, while individual trajectories are given as black lines.
}
\examples{
r <- runif(1000, 0, 1)
t <- seq(0, 5, by = 0.01)
plot_linear_rv_recovery(r, t, title = "Uniform distribution")

}
