% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocesses the data}
\usage{
preprocess_data(data, longgap)
}
\arguments{
\item{data}{The input data frame}

\item{longgap}{The value to use for the long gap in ms.}
}
\value{
The preprocessed data frame
}
\description{
This function preprocesses the input data by converting the trial column to numeric,
transforming the gap column values, and categorizing the itemtype based on serpos values.
}
