% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_or_load}
\alias{run_or_load}
\title{Execute an expression and save the result to a file or load the result from a file if it already exists.}
\usage{
run_or_load(expression, file, ..., force = FALSE)
}
\arguments{
\item{expression}{The expression to be evaluated or loaded.}

\item{file}{The file path where the result will be saved or loaded from.}

\item{...}{Additional arguments to be passed to the expression.}

\item{force}{Logical value indicating whether to force the evaluation of the expression, even if the file exists.}
}
\value{
The result of the expression.
}
\description{
This function allows you to either run an expression or load the result from a file.
}
\examples{
# Run an expression and save the result to a file
file <- tempfile(fileext = ".rds")
run_or_load(rnorm(1e7), file)
run_or_load(rnorm(1e7), file) # loads the result from the file

}
