% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{serial_recall}
\alias{serial_recall}
\title{Serial Recall Model}
\usage{
serial_recall(
  setsize,
  ISI = rep(0.5, setsize),
  item_in_ltm = rep(TRUE, setsize),
  prop = 0.2,
  prop_ltm = 0.5,
  tau = 0.15,
  gain = 25,
  rate = 0.1,
  prop_prim = 1,
  r_max = 1,
  lambda = 1,
  growth = "linear",
  ...
)
}
\arguments{
\item{setsize}{The number of items in the set.}

\item{ISI}{A numeric vector representing the inter-stimulus interval for each item.}

\item{item_in_ltm}{A logical vector indicating whether each item is in LTM.}

\item{prop}{The proportion of resources allocated to each item.}

\item{prop_ltm}{Proportion by which the resources used by LTM items are multiplied.}

\item{tau}{The threshold for recall probability.}

\item{gain}{The gain parameter for the recall probability function.}

\item{rate}{The rate at which resources are recovered.}

\item{prop_prim}{The output interference primacy parameter}

\item{r_max}{The maximum amount of resources.}

\item{lambda}{The exponent converting resources to strength.}

\item{growth}{The growth function for resource recovery. Either 'linear' or 'asy'.}
}
\value{
A numeric vector representing the recall probability for each item.
}
\description{
This function implements the model currently described in the draft on page 19.
It gives the predicted recall probability for each item in a set of items.
}
\details{
The function uses a simulation approach. It loops over trials
}
\examples{
serial_recall(setsize = 3, ISI = rep(0.5, 3))
}
